server/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "xince-server"
version = "0.1.0"
description = "Bootstrap API for XinCe"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13,<2.0",
  "aiosqlite>=0.20,<1.0",
  "asyncpg>=0.29,<1.0",
  "beautifulsoup4>=4.12,<5.0",
  "fastapi>=0.115,<1.0",
  "httpx>=0.27,<1.0",
  "pydantic>=2.7,<3.0",
  "pydantic-settings>=2.2,<3.0",
  "python-docx>=1.1,<2.0",
  "PyYAML>=6.0,<7.0",
  "SQLAlchemy[asyncio]>=2.0,<3.0",
  "redis>=5.0,<6.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
